requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.13.0
pytest>=7.4.0
torch>=2.0.0
scipy>=1.10.0
jupyter>=1.0.0
